closing the widget was occurring before the assined value was being captured, fixed that
